Revert "MythTV: quiet compiler warnings"
This reverts commit e70a6c007d8a23e9a3690fd65f1bfd5ad1cb9027.

+++ libavcodec/golomb.h
@@ -481,7 +481,7 @@ static inline int get_ur_golomb_jpegls(GetBitContext *gb, int k, int limit,
     } else {
         int i;
         for (i = 0; i + MIN_CACHE_BITS <= limit && SHOW_UBITS(re, gb, MIN_CACHE_BITS) == 0; i += MIN_CACHE_BITS) {
-            if (gb->size_in_bits <= (int)re_index) {
+            if (gb->size_in_bits <= re_index) {
                 CLOSE_READER(re, gb);
                 return -1;
             }
